Move 'RemoteFeedItem' to its own file
